skills/markdown-to-pdf/requirements.txt
# Markdown to PDF 转换工具依赖

# 核心依赖
markdown>=3.5.0
pymdown-extensions>=10.5
Pygments>=2.16.0
Pillow>=10.0.0
Jinja2>=3.1.0
beautifulsoup4>=4.12.0

# PDF生成
playwright>=1.40.0

# 可选依赖
# graphviz>=0.20.0  # 用于流程图生成
# weasyprint>=60.0  # 备选PDF生成方案

# 安装命令:
# pip install -r requirements.txt
# playwright install chromium
